Create new 'driver-extras' Maven module.
This commit creates the Maven module that will host goodies for the driver,
such as extra codecs. Note that Java >= 8 is required to compile this module.
It also updates Maven dependencies to more recent versions.

It also updates the Maven clirr plugin configuration with a workaround that
makes it compatible with Java 8.
The bug is actually in the BCEL library, see https://issues.apache.org/jira/browse/BCEL-173.
The workaround itself has been suggested here: https://github.com/RichardWarburton/lambda-behave/issues/31#issuecomment-86052095.

It also adds Maven animal-sniffer plugin to verify that compiled classes remain compatible with the driver's target API (Java 6).
Execution of the plugin is disabled for the new driver-extras module has this module currently requires Java 8, but there is no signature available yet for this JDK.

Finally, it also fixes wrong imports in MapperEnumTest. The file was importing jUnit's Assert class instead of TestNG's.